+++ run/workers/priorities.js
@@ -38,8 +38,10 @@ const priorities = {
     'low': [
         'processUser',
         'enforceDataRetentionForWorkspace',
+        'enrichProspect',
         'integrityCheckStarter',
         'integrityCheck',
+        'processDetectedProspects',
         'rpcHealthCheck',
         'rpcHealthCheckStarter',
         'explorerSyncCheck',
@@ -64,7 +66,7 @@ const priorities = {
 
 if (!isSelfHosted()) {
     priorities.high.push('updateApproximatedRecord', 'increaseStripeBillingQuota');
-    priorities.low.push('sendDemoExplorerLink', 'sendDripEmail', 'processDripEmails', 'enrichDemoProfile');
+    priorities.low.push('sendDemoExplorerLink', 'sendDripEmail', 'processDripEmails', 'enrichDemoProfile', 'processDetectedProspects', 'enrichProspect');
 }
 
 module.exports = priorities;